+++ RELEASE_NOTES.rst
@@ -2,6 +2,13 @@
 Release notes for bcolz
 =======================
 
+Changes from 0.11.2 to 0.11.3
+=============================
+
+- Implement feature #255 bcolz.zeros can create new ctables too, either
+  empty or filled with zeros. (#256 @FrancescElies @FrancescAlted)
+
+
 Changes from 0.11.1 to 0.11.2
 =============================